pyproject.toml
[project]
name = "bootdev-ai-agent"
version = "0.1.0"
description = "Simple Agentic AI tool using Google Gen AI SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.46.0",
    "python-dotenv>=1.2.1",
]
